data-pipeline/requirements.txt
azure-functions
azure-storage
numpy
pandas
requests
chardet
beautifulsoup4
lxml